requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.1
django-allauth==0.60.1
django-ckeditor==6.7.0
django-countries==7.5.1
django-crispy-forms==2.1
django-filter==23.5
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
idna==3.6
install==1.3.5
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
